+++ dev-src/com/palletops/awaze/client_builder.clj
@@ -6,7 +6,7 @@
    [clojure.tools.logging :refer [infof debugf]]
    [fipp.edn :refer [pprint]]
    [com.palletops.awaze.common
-    :refer [aws aws-client aws-client-factory aws-package?
+    :refer [aws aws-client aws-client-factory aws-ns-kw aws-package?
             camel->dashed camel->keyword
             coerce-value-form enum?
             to-data to-date type->symbol]])
@@ -28,37 +28,55 @@
   [^ParameterizedType type value]
   (let [raw (.getRawType type)
         [actual & others] (.getActualTypeArguments type)]
-    (cond
-     (or (= raw java.util.List) ((set (ancestors raw)) java.util.List))
-     (let [arg (gensym "arg")]
-       `(map (fn [~arg] ~(coerce-value-form actual arg)) ~value))
+    (with-meta
+      (cond
+       (or (= raw java.util.List) ((set (ancestors raw)) java.util.List))
+       (let [arg (gensym "arg")]
+         `(map (fn [~arg] ~(coerce-value-form actual arg)) ~value))
 
-     (or (= raw java.util.Map) ((set (ancestors raw)) java.util.Map))
-     (let [arg (gensym "arg")]
-       `(zipmap
-         (map (fn [~arg] ~(coerce-value-form actual arg)) (keys ~value))
-         (map
-          (fn [~arg] ~(coerce-value-form (first others) arg))
-          (vals ~value))))
+       (or (= raw java.util.Map) ((set (ancestors raw)) java.util.Map))
+       (let [arg (gensym "arg")]
+         `(zipmap
+           (map (fn [~arg] ~(coerce-value-form actual arg)) (keys ~value))
+           (map
+            (fn [~arg] ~(coerce-value-form (first others) arg))
+            (vals ~value))))
 
-     (or (= raw java.util.Collection)
-         ((set (ancestors raw)) java.util.Collection))
-     (let [arg (gensym "arg")]
-       `(map (fn [~arg] ~(coerce-value-form actual arg)) ~value))
+       (or (= raw java.util.Collection)
+           ((set (ancestors raw)) java.util.Collection))
+       (let [arg (gensym "arg")]
+         `(map (fn [~arg] ~(coerce-value-form actual arg)) ~value))
 
-     ;; (or (= raw com.amazonaws.services.ec2.model.DryRunSupportedRequest)
-     ;;     ((set (ancestors raw))
-     ;;      com.amazonaws.services.ec2.model.DryRunSupportedRequest))
-     ;; `(assert false
-     ;;          "Don't know how to hnadle args with type DryRunSupportedRequest")`
+       ;; (or (= raw com.amazonaws.services.ec2.model.DryRunSupportedRequest)
+       ;;     ((set (ancestors raw))
+       ;;      com.amazonaws.services.ec2.model.DryRunSupportedRequest))
+       ;; `(assert false
+       ;;   "Don't know how to hnadle args with type DryRunSupportedRequest")`
 
-     :else
-     (assert
-      nil
-      (str "Don't know how to handle generic args with raw type "
-           raw " "(class raw))))))
+       :else
+       (assert
+        nil
+        (str "Don't know how to handle generic args with raw type "
+             raw " "(class raw))))
+      {:tag raw})))
 
 ;;; # Bean Factories
+(defn- select-vararg-overload
+  [setters]
+  (or (first
+       (remove
+        #(.isArray ^Class (first (.getParameterTypes ^Method %)))
+        setters))
+      (first setters)))
+
+(defn- filter-vararg-overloads
+  "Filter out overloaded setters that just provide convenience
+  functions to other setters."
+  [setters]
+  (->> (group-by #(.getName ^Method %) setters)
+       (map second)
+       (map select-vararg-overload)))
+
 (defn- setters
   "Return a sequence of vectors with the hyphenated keywords and setter method
   for the specified `class`."
@@ -67,6 +85,7 @@
        (filterv public-method?)
        (filterv #(.startsWith (.getName ^Method %) "set"))
        (filterv #(= 1 (count (.getParameterTypes ^Method %))))
+       filter-vararg-overloads
        (mapv #(vector (camel->keyword (subs (.getName ^Method %) 3)) %))))
 
 (def primitive-default-value
@@ -85,23 +104,40 @@
                                       (.getConstructors class)
                                       (sort-by sortf)
                                       (partition-by sortf)
-                                      first)
-        arglist (.getParameterTypes ctor)]
-    (if (seq others)
-      ;; need to give the compiler type info to pick the correct overload
-      [ctor
-       (mapv
-        #(let [arg (with-meta (gensym "arg")
-                     {:tag (symbol (.getName ^Class %))})
-               val (if (aws-package? %)
-                     `(~(type->symbol %) {})
-                     (primitive-default-value %))]
-           [arg val])
-        arglist)]
-      ;; single overload - just return nils or primitive values
-      [ctor (->> arglist
-                 (map primitive-default-value)
-                 (map #(vector (gensym "arg") %)))])))
+                                      first)]
+    (if-let [arglist (and ctor (.getParameterTypes ctor))]
+      (if (seq others)
+        ;; need to give the compiler type info to pick the correct overload
+        [ctor
+         (mapv
+          #(let [arg (with-meta (gensym "arg")
+                       {:tag (symbol (.getName ^Class %))})
+                 val (if (aws-package? %)
+                       `(~(type->symbol %) {})
+                       (primitive-default-value %))]
+             [arg val])
+          arglist)]
+        ;; single overload - just return nils or primitive values
+        [ctor (->> arglist
+                   (map primitive-default-value)
+                   (map #(vector (gensym "arg") %)))]))))
+
+(def static-constructors
+  "A map from class name to static constructor functions, for classes that
+  have no constructors."
+  {"com.amazonaws.regions.Region"
+   {:static "getRegion"
+    :arg-bean com.amazonaws.regions.Regions}})
+
+(defn has-from-value? [^Class class]
+  (try (let [v (.getMethod class "fromValue" (into-array Class [String]))]
+         (bit-and (.getModifiers v) Modifier/PUBLIC))
+       (catch NoSuchMethodException _)))
+
+(defn has-from-name? [^Class class]
+  (try (let [v (.getMethod class "fromName" (into-array Class [String]))]
+         (bit-and (.getModifiers v) Modifier/PUBLIC))
+       (catch NoSuchMethodException _)))
 
 (defn bean-factory
   "Return a form defining a factory for the specified bean `class`, based on a
@@ -109,24 +145,32 @@
   [^Class class]
   (let [bean (gensym "bean")
         m (gensym "m")
-        v (gensym "v")]
+        v (gensym "v")
+        class-name (.getName class)]
     `(defn ~(with-meta (type->symbol class) {:tag class})
        [~m]
        ~(cond
          (.isInterface class)
          `(throw (ex-info "Trying to construct an interface"
-                          {:type ~(.getName class)
+                          {:type ~class-name
                            :value ~m}))
 
          (enum? class)
-         (if (= class com.amazonaws.services.ec2.model.InstanceType)
-           `(. ~class ~'fromValue (name ~m))
-           `(Enum/valueOf ~class (name ~m)))
+         (cond
+          (has-from-value? class) `(. ~class ~'fromValue (name ~m))
+          (has-from-name? class) `(. ~class ~'fromName (name ~m))
+          :else `(Enum/valueOf ~class (name ~m)))
+
+         (static-constructors class-name)
+         (let [s (static-constructors class-name)]
+           `(~(symbol class-name (:static s))
+             ;; TODO, fix this in a better manner
+             ~(coerce-value-form (:arg-bean s) m)))
 
          :else
          (let [[ctor args] (required-constructor-args class)]
            `(let [~@(apply concat args)
-                  ~bean (new ~(symbol (.getName class))
+                  ~bean (new ~(symbol class-name)
                              ~@(map first args))]
               ~@(for [[kw ^Method method] (setters class)
                       :let [args (.getGenericParameterTypes method)]
@@ -139,7 +183,6 @@
                           arg-type v))))
               ~bean))))))
 
-
 (defn bean-instance
   "Return an instance of the specified bean.  If a value can't be constructed,
   return nil?."
@@ -148,16 +191,19 @@
    (.isInterface class) nil
    (enum? class) nil
    :else
-   (let [[^Constructor ctor args] (required-constructor-args class)]
-     (.newInstance
-      ctor
-      (into-array Object
-                  (->> args
-                       (map second)
-                       (map (fn [x]     ; remove bean constructor expressions
-                              (if (or (list? x) (instance? clojure.lang.Cons x))
-                                nil
-                                x)))))))))
+   (if-let [[^Constructor ctor args] (required-constructor-args class)]
+     (if ctor
+       (let [vargs (seq
+                    (->>
+                     args
+                     (map second)
+                     (map (fn [x] ; remove bean constructor expressions
+                            (if (or (list? x) (instance? clojure.lang.Cons x))
+                              nil
+                              x)))))]
+         (.newInstance
+          ctor
+          (into-array Object vargs)))))))
 
 (defmulti setter-arg-type
   "Return a sequence of bean types used in an argument"
@@ -214,12 +260,21 @@
                     (map second)
                     aws-bean-types))))))
 
-(defn request-bean-factories
-  "Return a sequence of bean factory definitions, for all beans used as
-  arguments to the specified `methods`, a map from keyword to Method sequence."
+(defn request-beans-by-ns
+  "Return a map with a sequence of beans for each
+  top level namespace"
   [methods]
   (->> (request-beans methods)
-       (map bean-factory)))
+       (group-by aws-ns-kw)))
+
+(defn request-bean-factories
+  "Return a map with a sequence of bean factory definitions for each
+  top level namespace, for all beans used as arguments to the
+  specified `methods`, a map from keyword to Method sequence."
+  [beans]
+  (infof "r-b-f %s" (pr-str beans))
+  (flush)
+  (mapv bean-factory beans))
 
 ;;; ## Client Factory
 (defn client-factory
@@ -264,11 +319,15 @@
   [types]
   (map
    (fn [^Class type n]
+     (debugf "arg-names type %s %s" (pr-str type) (aws-package? type))
      (let [s (symbol
               (str (last (split (name (camel->keyword (.getName type))) #"\."))
                    "-" n))]
        (if (aws-package? type)
-         (if-let [instance (bean-instance type)]
+         (if-let [instance (and
+                            (not= type
+                                  com.amazonaws.metrics.RequestMetricCollector)
+                            (bean-instance type))]
            (do
              (debugf "instance type %s" (pr-str type))
              `{:keys [~@(try
@@ -278,9 +337,14 @@
                            (comp symbol name)
                            (keys (dissoc (to-data instance)
                                          :request-credentials
-                                         :request-client-options)))
-                          (catch java.lang.reflect.InvocationTargetException _))]})
-           s)
+                                         :request-client-options
+                                         :request-metric-collector
+                                         :dry-run-request)))
+                          (catch java.lang.reflect.InvocationTargetException
+                              _))]})
+           (do
+             (debugf "instance type %s no bean-instance" (pr-str type))
+             s))
          s)))
    types (range)))
 
@@ -420,17 +484,22 @@
   [client-kw client-class]
   (let [class (Class/forName (name client-class))
         sym (symbol (name client-kw))
-        methods (client-methods class)]
-    `(do
-       ~(client-factory client-kw client-class)
-       ~@(request-bean-factories methods)
-       (defmulti ~sym (fn ~sym [m#] (:fn m#)))
-       ~@(for [[fn-kw methods] methods]
-           `(do
-              (defmethod ~(symbol (name client-kw)) ~fn-kw
-                ~@(client-method client-kw client-class methods))
-              ~(client-map-fn client-kw client-class fn-kw methods)
-              ~(client-fn client-kw client-class fn-kw methods))))))
+        methods (client-methods class)
+        ns-beans (request-beans-by-ns methods)
+        beans (client-kw ns-beans)]
+    (infof "bean ns %s" (pr-str (keys ns-beans)))
+    (infof "beans %s" (pr-str beans))
+    [`(do
+        ~(client-factory client-kw client-class)
+        ~@(request-bean-factories beans)
+        (defmulti ~sym (fn ~sym [m#] (:fn m#)))
+        ~@(for [[fn-kw methods] methods]
+            `(do
+               (defmethod ~(symbol (name client-kw)) ~fn-kw
+                 ~@(client-method client-kw client-class methods))
+               ~(client-map-fn client-kw client-class fn-kw methods)
+               ~(client-fn client-kw client-class fn-kw methods))))
+     (dissoc ns-beans client-kw)]))
 
 (def apis
   {:ec2 'com.amazonaws.services.ec2.AmazonEC2Client
@@ -449,7 +518,7 @@
    :elasticbeanstalk 'com.amazonaws.services.elasticbeanstalk.AWSElasticBeanstalkClient
    :elasticloadbalancing 'com.amazonaws.services.elasticloadbalancing.AmazonElasticLoadBalancingClient
    :elasticmapreduce 'com.amazonaws.services.elasticmapreduce.AmazonElasticMapReduceClient
-   :glacier 'com.amazonaws.auth.DefaultAWSCredentialsProviderChain
+   :glacier 'com.amazonaws.services.glacier.AmazonGlacierClient
    ;; :opsworks 'com.amazonaws.services.opsworks.AWSOpsWorksClient
    :rds 'com.amazonaws.services.rds.AmazonRDSClient
    :redshift 'com.amazonaws.services.redshift.AmazonRedshiftClient
@@ -458,42 +527,88 @@
    :simpledb 'com.amazonaws.services.simpledb.AmazonSimpleDBClient
    :simpleemail 'com.amazonaws.services.simpleemail.AmazonSimpleEmailServiceClient
    :sns 'com.amazonaws.services.sns.AmazonSNSClient
-   :sqs 'com.amazonaws.auth.DefaultAWSCredentialsProviderChain
+   :sqs 'com.amazonaws.services.sqs.AmazonSQSClient
    :storagegateway 'com.amazonaws.services.storagegateway.AWSStorageGatewayClient})
 
 (def api-refer-exclude
   "Map of clojure symbols to exclude for each api namespace to avoid
   redefinition warnings."
-  {:ec2 '[filter]})
+  {:ec2 '[filter]
+   :autoscaling '[filter]
+   :rds '[filter]})
 
 (defn api-ns-form
   "Return a `ns` form for the given api."
-  [pp api]
-  (with-out-str
-    (pp
-     `(~'ns ~(symbol (str "com.palletops.awaze." (name api)))
-        ~@(if-let [excludes (api api-refer-exclude)]
-            [`(:refer-clojure :exclude ~excludes)])
-        (:require [com.palletops.awaze.common])))))
+  [pp api ks & segs]
+  (let [s (symbol (str "com.palletops.awaze."
+                       (string/join "." segs)
+                       (if (seq segs) ".")
+                       (name api)))]
+    (with-out-str
+      (pp
+       `(~'ns ~s
+          ~@(if-let [excludes (api api-refer-exclude)]
+              [`(:refer-clojure :exclude ~excludes)])
+          (:require
+           [com.palletops.awaze.common]
+           [~s :as ~(symbol (name api))]
+           ~@(map (fn [k]
+                    [(symbol (str "com.palletops.awaze.beans." (name k)))
+                     :as (symbol (name k))])
+                  ks)))))))
 
 (defn gen-api
-  [target api class & {:keys [pretty-print]}]
+  [target api class {:keys [pretty-print]}]
   (println "Generating" api "in" (str target))
-  (let [pp (if pretty-print pprint identity)]
+  (let [pp (if pretty-print pprint identity)
+        [form beans] (add-aws-client api class)]
     (binding [*print-meta* true]
       (spit target
-            (str (api-ns-form pp api)
+            (str (api-ns-form pp api (keys beans))
                  \newline \newline
-                 (with-out-str (pp (add-aws-client api class))))))))
+                 (with-out-str (pp form)))))
+    beans))
 
 (defn gen-apis
-  [target & {:keys [pretty-print]}]
-  (doseq [[api class] apis]
-    (gen-api
-     (doto (file target "com" "palletops" "awaze"
-                 (str (name api) ".clj"))
+  [target {:keys [pretty-print]}]
+  (reduce
+   (fn [m [api class]]
+     (merge-with
+      concat
+      m
+      (gen-api
+       (doto (file target "com" "palletops" "awaze"
+                   (str (name api) ".clj"))
+         (-> (.getParentFile) (.mkdirs)))
+       api class {:pretty-print pretty-print})))
+   {}
+   apis))
+
+(defn gen-bean-ns
+  "Generate a bean namespace"
+  [target kw beans {:keys [pretty-print]}]
+  (infof "target %s count %s pp %s" target (count beans) (pr-str pretty-print))
+  (infof "beans %s" (pr-str beans))
+  (let [pp (if pretty-print pprint println)
+        form (str (api-ns-form pp kw nil "beans")
+                  \newline \newline
+                  (with-out-str
+                    (pp `(do ~@(request-bean-factories beans)))))]
+    (binding [*print-meta* true]
+      (spit target form))))
+
+(defn gen-beans
+  "Generate bean namespaces"
+  [gen-src ns-beans {:keys [pretty-print]}]
+  (infof "gen-beans %s" ns-beans)
+  (doseq [[kw beans] ns-beans]
+    (println "Generating bean namespace" kw "with" (count beans) "beans")
+    (gen-bean-ns
+     (doto (file gen-src "com" "palletops" "awaze" "beans"
+                 (str (name kw) ".clj"))
        (-> (.getParentFile) (.mkdirs)))
-     api class :pretty-print pretty-print)))
+     kw beans
+     {:pretty-print pretty-print})))
 
 (defn -main
   [& [pp]]
@@ -501,4 +616,14 @@
     (let [gen-src (file target-path "generated")]
       (println "Generating awaze source to" (str gen-src))
       (.mkdirs gen-src)
-      (gen-apis gen-src :pretty-print pp))))
+      (binding [*print-meta* true]
+        (let [ns-beans (gen-apis gen-src {:pretty-print pp})
+              ns-beans (reduce
+                        (fn [m b]
+                          (update-in m [(aws-ns-kw b)]
+                                     #(concat [b] %)))
+                        ns-beans
+                        (map :arg-bean (vals static-constructors)))]
+          (println "Generating bean namespaces for"
+                   (count ns-beans) "namespaces")
+          (gen-beans gen-src ns-beans {:pretty-print pp}))))))