+++ test/com/palletops/awaze/ec2_test.clj
@@ -4,4 +4,5 @@
    [com.palletops.awaze.ec2 :refer [describe-instances-map]]))
 
 (deftest simple-test
-  (is (map? (describe-instances-map {:access-key "a" :secret-key "b"}))))
+  (is (map? (describe-instances-map {:access-key "a" :secret-key "b"})))
+  (is (= 2 (count (-> #'describe-instances-map meta :arglists)))))