+++ ReleaseNotes.md
@@ -1,3 +1,14 @@
+## 0.1.2
+
+- Update to AWS SDK 1.8.3
+  AWS SDK 1.8.3 includes the t2.small and t2.medium instance types
+
+- Fix glacier, sqs, autoscaling and rds
+
+- Remove common keys from request object arglists
+  Removes the :dry-run-request and :request-metric-collector keys from the
+  arglist
+
 ## 0.1.1
 
 - Fix runtime coercions